+++ src/LogSearchShipper/LogSearchShipper.csproj
@@ -67,7 +67,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Appenders\MainLogAppender.cs" />
+    <Compile Include="Log4net\MainLogAppender.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
